refactor(examples): move ClientRouter to apps/polls/urls/client_router.rs
Per the project URL routing layout convention, the client-side router for
the polls SPA must live under the polls app at
`apps/polls/urls/client_router.rs`, not as a flat `client/router.rs` file.

Changes:

- Create `apps/polls/urls/` directory with two submodules:
  - `server_urls.rs`: existing `routes() -> ServerRouter` (was `urls.rs`)
  - `client_router.rs`: ClientRouter setup, init_global_router, with_router
    (was `client/router.rs`)
- Rewrite `apps/polls/urls.rs` as a cfg-gated aggregator that re-exports
  `routes` (native) and `init_global_router`/`with_router` (wasm).
- Drop `#[cfg(native)]` from `pub mod apps;` in `lib.rs` because `apps`
  now hosts WASM submodules; cfg-gate at the leaf instead
  (`apps/polls.rs` gates models/serializers/views/PollsConfig as native).
- Update `src/client.rs` to remove `pub mod router;`.
- Update `src/client/lib.rs` to import from `crate::apps::polls::urls::client_router`.
- Use `reinhardt::ClientRouter` (canonical top-level facade path), not
  `reinhardt::urls::ClientRouter` (not re-exported via urls on rc.28).

`config/urls.rs` is unchanged: it still calls
`crate::apps::polls::urls::routes()` which is re-exported from the aggregator
on native.

Verified:
- `cargo check -p examples-tutorial-basis` (native): clean (17.47s)
- `cargo check -p examples-tutorial-basis --lib --target wasm32-unknown-unknown`: clean (0.81s)

Fixes #4274
Refs #4234
