+++ examples/examples-tutorial-basis/src/apps/polls.rs
@@ -6,12 +6,17 @@
 //! - Forms and generic views
 //! - Admin panel integration
 
+#[cfg(native)]
 use reinhardt::app_config;
 
+#[cfg(native)]
 pub mod models;
+#[cfg(native)]
 pub mod serializers;
 pub mod urls;
+#[cfg(native)]
 pub mod views;
 
+#[cfg(native)]
 #[app_config(name = "polls", label = "polls")]
 pub struct PollsConfig;
