+++ examples/examples-tutorial-basis/src/apps/polls/urls/server_urls.rs
@@ -1,0 +1,16 @@
+//! Server-side URL configuration for the polls application.
+
+use reinhardt::ServerRouter;
+
+use crate::apps::polls::views;
+
+// Note: This function is called by config/urls.rs via .mount("/polls/", ...).
+// Do NOT add #[routes] here - that would create a duplicate registration
+// without the mount prefix.
+pub fn routes() -> ServerRouter {
+	ServerRouter::new()
+		.endpoint(views::index)
+		.endpoint(views::detail)
+		.endpoint(views::results)
+		.endpoint(views::vote)
+}
