+++ examples/examples-tutorial-basis/src/client.rs
@@ -1,10 +1,9 @@
 //! Client-side code (WASM)
 //!
 //! This module contains all client-side code that runs in the browser.
+//! Client-side routing lives under `crate::apps::polls::urls::client_router`.
 
 pub mod lib;
-
-pub mod router;
 
 pub mod pages;
